requirements.txt
numpy==1.24.3
pandas==2.0.1
streamlit==1.22.0
openai==0.27.6
requests==2.30.0
pyYAML==6.0
sqlalchemy==2.0.15